admin auth

+++ nodereactroot/server/index.js
@@ -18,26 +18,51 @@ app.set('view engine', 'ejs');
 app.use(cors())
 app.use(express.json())
 app.use('/api',router)
-app.get('/admin', async function(req, res) {
-    var p = await models.Person.findAll()
-    var people = []
-    p.map(person=>people.push(person.dataValues))
-
-    //const { html } = new Jsontableify().toHtml(l)
-    res.render('pages/admin', {
-        people: people,
-    });
-});
+app.get('/admin', async function(req,res){
+    res.render('pages/auth');
+})
+//app.get('/admin', async function(req, res) {
+//    var p = await models.Person.findAll()
+//    var people = []
+//    p.map(person=>people.push(person.dataValues))
+//
+//    //const { html } = new Jsontableify().toHtml(l)
+//    res.render('pages/admin', {
+//        people: people,
+//    });
+//});
 app.post('/admin',async function(req,res){
-    let toDel = req.body.id.split(' ')
-    delPeople = await models.Person.findAll()
-    models.Person.destroy({
-        where: {
-            id:toDel
+    console.log(req.body)
+    if (req.body.password==process.env.ADMIN_PASS){
+        var p = await models.Person.findAll()
+        var people = []
+        p.map(person=>people.push(person.dataValues))
+
+        //const { html } = new Jsontableify().toHtml(l)
+        res.render('pages/admin', {
+            people: people,
+        });
+    }else{
+        if (req.body.id){
+            let toDel = req.body.id.split(' ')
+            delPeople = await models.Person.findAll()
+            models.Person.destroy({
+                where: {
+                    id:toDel
+                }
+            })
+            let answer = 'Удалено!'
+            res.render('pages/template',{
+                    answer:answer
+            })
+        }else{
+            let answer = 'wrong password!'
+            res.render('pages/template',{
+                    answer:answer
+            })
         }
-    })
-    res.end('Successfully deleted!')
-});
+    }
+}); 
 
 app.use(errorHandler) //должен стоять в конце списка use
